+++ lib/gradient_containers_plus.dart
@@ -9,4 +9,5 @@ export 'gradient_containers_plus/containers/mesh_gradient_container.dart';
 export 'gradient_containers_plus/containers/neon_gradient_container.dart';
 export 'gradient_containers_plus/containers/radial_gradient_container.dart';
 export 'gradient_containers_plus/containers/sweep_gradient_container.dart';
+export 'gradient_containers_plus/containers/liquid_press_container.dart';
 export 'gradient_containers_plus/presets/gradient_presets.dart';